build-and-push-assets.md // update documentation example to always run when creating a tag.
The current example in `Simple usage example:` only runs when some specific file are changing. But when creating a Tag (and Release) we always want to execute the workflow and build assets in a minified and optimized version. Therfor we need to add `on.push.tags` with `'*'` to ensure the workflow is always triggered.