build-and-push-assets.md // add on.push.branches
add missing `on.push.branches` with `['*']` to example workflow. When only defining `on.push.tags` the branches will not trigger any workflow, eventho we have `on.push.paths` defined.

When we define both `on.push.branches` and `on.push.paths`, the workflow will only be dispatched when both filters are satisfied.