build-assets-compilation // set ENCORE_ENV to deprecated and introduce new BUILD_ENV variable.
As described in #33 the "ENCORE_ENV"-variable is way to specific and relates to [Symfony Webpack Encore](https://github.com/symfony/webpack-encore). 
Acutally it is possible to use differen build tools in this workflow through [Composer Assets Compiler](https://github.com/inpsyde/composer-asset-compiler). This is why we will set the "ENCORE_ENV" to deprecated and introduce a new env variable "BUILD_ENV".

Signed-off-by: Christian Leucht <3417446+Chrico@users.noreply.github.com>