Feature/setup node cache mode correctly and generalize inputs (#57)
This PR does generalize the `on.workflow_call.inputs` and checks, based on the `PACKAGE_MANAGER`, if a .lock-file is present. If so, the cache is used in `actions/setup-node@v3`. 
Additionally, for all reusable workflows which do not use `composer compile-assets`, we make use of either `npm ci` or `yarn --frozen-lockfile` to speed up the installation in case we found a .lock-file.