Feature - deprecate inputs.PHP_MATRIX and introduce inputs.PHP_VERSION (#69)
The ´inputs.PHP_MATRIX´ is being deprecated and a new ´inputs.PHP_VERSION´ is added to be consistent across all php-based reusable workflows. 
The calling workflow can now use a `strategy.matrix` to put the right php version into that `inputs.PHP_VERSION` and run the given reusable workflow in a matrix. We still support the `inputs.PHP_MATRIX` and will print a deprecated message until it will be removed at a later point.